Create CourseInfoTabInfoViewModelAdapter
