+++ Stepic/Modules/CourseInfoTabInfo/CourseInfoTabInfoPresenter.swift
@@ -19,142 +19,11 @@ final class CourseInfoTabInfoPresenter: CourseInfoTabInfoPresenterProtocol {
         var viewModel: CourseInfoTabInfo.ShowInfo.ViewModel
 
         if let course = response.course {
-            viewModel = .init(state: .result(data: self.courseToViewModel(course: course)))
+            viewModel = .init(state: .result(data: .init(course: course)))
         } else {
             viewModel = .init(state: .loading)
         }
 
         self.viewController?.displayCourseInfo(viewModel: viewModel)
-    }
-
-    // MARK: Prepare view data
-
-    private func courseToViewModel(course: Course) -> CourseInfoTabInfoViewModel {
-        let authorText = self.formattedAuthor(authors: course.authors)
-        let timeToCompleteText = self.formattedTimeToComplete(seconds: course.timeToComplete)
-        let languageText = self.localizedLanguage(code: course.languageCode)
-
-        let certificateText = self.formattedCertificate(course: course)
-        let certificateDetailsText = self.formattedCertificateDetails(
-            conditionPoints: course.certificateRegularThreshold,
-            distinctionPoints: course.certificateDistinctionThreshold
-        )
-
-        let instructorsViewModel = course.instructors.map { user in
-            CourseInfoTabInfoInstructorViewModel(
-                avatarImageURL: URL(string: user.avatarURL),
-                title: "\(user.firstName) \(user.lastName)",
-                description: user.bio
-            )
-        }
-
-        let actionButtonTitle = course.enrolled
-            ? NSLocalizedString("ContinueLearning", comment: "")
-            : NSLocalizedString("JoinCourse", comment: "")
-
-        return CourseInfoTabInfoViewModel(
-            author: authorText,
-            introVideoURL: self.getIntroVideoURL(course: course),
-            introVideoThumbnailURL: URL(string: course.introVideo?.thumbnailURL ?? ""),
-            aboutText: course.summary.trimmingCharacters(in: .whitespaces),
-            requirementsText: course.requirements.trimmingCharacters(in: .whitespaces),
-            targetAudienceText: course.audience.trimmingCharacters(in: .whitespaces),
-            timeToCompleteText: timeToCompleteText,
-            languageText: languageText,
-            certificateText: certificateText,
-            certificateDetailsText: certificateDetailsText,
-            instructors: instructorsViewModel,
-            actionButtonTitle: actionButtonTitle
-        )
-    }
-
-    private func getIntroVideoURL(course: Course) -> URL? {
-        if let introVideo = course.introVideo, !introVideo.urls.isEmpty {
-            return introVideo.getUrlForQuality(VideosInfo.watchingVideoQuality)
-        } else {
-            return URL(string: course.introURL)
-        }
-    }
-
-    private func formattedAuthor(authors: [User]) -> String {
-        if authors.isEmpty {
-            return ""
-        } else {
-            var authorString = authors.reduce("") { result, user in
-                result + "\(user.firstName) \(user.lastName), "
-            }.trimmingCharacters(in: .whitespaces)
-            authorString.removeLast()
-
-            return authorString
-        }
-    }
-
-    private func formattedTimeToComplete(seconds: Int?) -> String {
-        if let seconds = seconds {
-            let hour = 3600.0
-            let hours = Int(ceil(Double(seconds) / hour))
-
-            let pluralizedHoursString = StringHelper.pluralize(
-                number: hours,
-                forms: [
-                    NSLocalizedString("hours1", comment: ""),
-                    NSLocalizedString("hours234", comment: ""),
-                    NSLocalizedString("hours567890", comment: "")
-                ]
-            )
-
-            return "\(hours) \(pluralizedHoursString)"
-        } else {
-            return ""
-        }
-    }
-
-    private func localizedLanguage(code: String) -> String {
-        return Locale.current.localizedString(forLanguageCode: code)?.capitalized ?? ""
-    }
-
-    private func formattedCertificate(course: Course) -> String {
-        let certificateText = course.certificate.trimmingCharacters(in: .whitespaces)
-        if certificateText.isEmpty {
-            return course.certificateRegularThreshold ?? 0 > 0 && course.certificateDistinctionThreshold ?? 0 > 0
-                ? NSLocalizedString("Yes", comment: "")
-                : NSLocalizedString("No", comment: "")
-        } else {
-            return certificateText
-        }
-    }
-
-    private func formattedCertificateDetails(
-        conditionPoints: Int?,
-        distinctionPoints: Int?
-    ) -> String {
-        let formattedCondition = self.formattedCertificateDetailTitle(
-            NSLocalizedString("CertificateCondition", comment: ""),
-            points: conditionPoints
-        )
-        let formattedDistinction = self.formattedCertificateDetailTitle(
-            NSLocalizedString("WithDistinction", comment: ""),
-            points: distinctionPoints
-        )
-
-        return "\(formattedCondition)\n\(formattedDistinction)".trimmingCharacters(
-            in: .whitespacesAndNewlines
-        )
-    }
-
-    private func formattedCertificateDetailTitle(_ title: String, points: Int?) -> String {
-        if let points = points, points > 0 {
-            let pluralizedPointsString = StringHelper.pluralize(
-                number: points,
-                forms: [
-                    NSLocalizedString("points1", comment: ""),
-                    NSLocalizedString("points234", comment: ""),
-                    NSLocalizedString("points567890", comment: "")
-                ]
-            )
-            return "\(title): \(points) \(pluralizedPointsString)"
-        } else {
-            return ""
-        }
     }
 }
